requirements.txt
fastapi[all]==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.16.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
weasyprint==61.2
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9  # Rate limiting

